refactor(api): remove dead transaction middleware and rename database provider
The TransactionMiddleware, TRANSACTION symbol, and TransactionProvider were
completely unused. Strip them out and move the remaining DatabaseProvider from
common/middleware/db-transaction-middleware.ts to common/database-provider.ts.
