+++ apps/api/src/common/database-provider.ts
@@ -1,0 +1,9 @@
+import { Provider, Scope } from "@nestjs/common";
+import { getDb } from "@tambo-ai-cloud/db";
+
+export const DATABASE = Symbol("DATABASE");
+export const DatabaseProvider: Provider = {
+  provide: DATABASE,
+  scope: Scope.REQUEST,
+  useFactory: () => getDb(process.env.DATABASE_URL!),
+};
