+++ apps/api/src/threads/threads.service.test.ts
@@ -19,7 +19,7 @@ import {
   createMockDBProject,
   createMockDBThread,
 } from "@tambo-ai-cloud/testing";
-import { DATABASE } from "../common/middleware/db-transaction-middleware";
+import { DATABASE } from "../common/database-provider";
 import { AnalyticsService } from "../common/services/analytics.service";
 import { AuthService } from "../common/services/auth.service";
 import { EmailService } from "../common/services/email.service";
